+++ src/main/java/woowacourse/shoppingcart/domain/cart/CartItem.java
@@ -1,7 +1,8 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.cart;
 
 import java.util.Objects;
 
+import woowacourse.shoppingcart.domain.Quantity;
 import woowacourse.shoppingcart.domain.product.Product;
 import woowacourse.shoppingcart.exception.DuplicateProductInCartException;
