+++ src/test/java/woowacourse/shoppingcart/dao/CartItemDaoTest.java
@@ -14,8 +14,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
 
-import woowacourse.shoppingcart.domain.CartItem;
 import woowacourse.shoppingcart.domain.Quantity;
+import woowacourse.shoppingcart.domain.cart.CartItem;
 import woowacourse.shoppingcart.domain.product.Product;
 import woowacourse.shoppingcart.domain.product.ProductStock;
 import woowacourse.shoppingcart.domain.product.ThumbnailImage;
