+++ src/test/java/woowacourse/shoppingcart/domain/customer/CustomerTest.java
@@ -1,10 +1,9 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.customer;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import woowacourse.shoppingcart.domain.customer.Customer;
 
 class CustomerTest {
 
@@ -13,5 +12,5 @@ class CustomerTest {
     void createCustomer() {
         assertDoesNotThrow(() -> new Customer("email@email.com", "password!2", "user"));
     }
-    
+
 }
